requirements.txt
Django==1.10.5
git+git://github.com/disqus/disqus-python.git
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6
django-extensions==1.7.6
textblob==0.12.0
pandas==0.19.2
simplejson
sklearn
nltk
pytz
django-graphos
django-picklefield
beautifulsoup4
scipy
django-inline-svg
django-rq

